requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.21.0
easyocr>=1.7.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
transformers>=4.30.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
albumentations>=1.3.0
tqdm>=4.60.0
requests>=2.25.0 